+++ CHANGELOG.md
@@ -20,7 +20,7 @@ Please make sure to add your changes to the appropriate categories:
 
 ### Added
 
-- Added `BiquadCascade` and `Butterworth` low, high, and band-pass biquad filters
+- Added `BiquadCascade` and `Butterworth` low-pass, high-pass, band-pass, and band-stop biquad filters
 
 ### Changed